+++ models/server.js
@@ -3,13 +3,13 @@ module.exports = function(app, hosts, servers){
 	return function(location, httpsOptions){
 		// get location and create route holder for app
 		var location = isNaN(location) ? location : 'http://localhost:'+location ;
-		app.location = require('url').parse(location);
+		app.location = typeof location == 'object' ? location : require('url').parse(location) ;
 		app.routes = { get: [], post: [], header: [], footer: [], missing: [], error: [] }
 		hosts[app.location.host] = app
 		
 		// get protocol and port
 		var protocol = app.location.protocol === 'http:' ? require('http') : require('https')
-		var port = app.location.protocol === 'http:' ? (app.location.port || 80) : 443 ;
+		var port = app.location.protocol === 'http:' ? (app.location.port || 80) : (app.location.port || 443) ;
 		
 		// create server
 		if(!servers[port]){
@@ -18,6 +18,7 @@ module.exports = function(app, hosts, servers){
 				? protocol.createServer(host).listen(port) 
 				: protocol.createServer(httpsOptions, host).listen(port) ;
 			servers[port] = server
+			
 		} else {
 			var server = servers[port]
 		}